refactor: reorganize deployCommands function and update import paths
