+++ bot/events/handlers/clientReady.ts
@@ -1,6 +1,6 @@
 import { ActivityType, Events } from 'discord.js';
 import type { Event } from '../event';
-import { deployCommands } from '../../commands/deploy-commands';
+import { deployCommands } from '../../deploy-commands';
 import { logger } from '../../..';
 import { birthdayCron } from '../../../cron/birthdayCron';
 import { statusCron } from '../../../cron/statusCron';
